azureappintegration/fortigate-azure-chatbot/requirements_windows_gpu.txt
# FortiGate Azure Chatbot - Windows GPU Requirements
# Optimized for ASUS ProArt + NVIDIA GPU

# Core PyTorch with CUDA support
--index-url https://download.pytorch.org/whl/cu118
torch>=2.1.0+cu118
torchvision>=0.16.0+cu118
torchaudio>=2.1.0+cu118

# Transformers and AI libraries
transformers[torch]>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
peft>=0.6.0
datasets>=2.14.0

# Quantum compression libraries
tensorly>=0.8.1
tensorly-torch>=0.3.0
qiskit>=0.45.0
qiskit-machine-learning>=0.7.0

# Streamlit and web interface
streamlit>=1.28.0
streamlit-audiorecorder>=0.0.5
streamlit-plotly-events>=0.0.6

# Visualization and data processing
plotly>=5.17.0
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Azure and cloud integration
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
azure-storage-blob>=12.19.0
azure-cognitiveservices-speech>=1.34.0

# OpenAI and LLM integration
openai>=1.3.0
langchain>=0.0.350
langchain-openai>=0.0.2
pinecone-client>=2.2.4

# Audio processing
speechrecognition>=3.10.0
pydub>=0.25.1
pyaudio>=0.2.11

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
psutil>=5.9.0
GPUtil>=1.4.0

# Development and testing
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0

# Windows-specific optimizations
pywin32>=306
wmi>=1.5.1
